refactored package name to use only lower case
Warning was being shown in android manifest
