+++ app/src/main/java/com/dsciitp/shabd/dictionary/DictionaryActivity.java
@@ -1,4 +1,4 @@
-package com.dsciitp.shabd.Dictionary;
+package com.dsciitp.shabd.dictionary;
 
 import android.net.Uri;
 import android.os.AsyncTask;
