+++ app/src/main/java/com/dsciitp/shabd/home/HomeFragment.java
@@ -1,4 +1,4 @@
-package com.dsciitp.shabd.Home;
+package com.dsciitp.shabd.home;
 
 import android.os.Bundle;
 import android.support.annotation.NonNull;
