+++ app/src/main/java/com/dsciitp/shabd/quickActions/QuickActionFragment.java
@@ -1,4 +1,4 @@
-package com.dsciitp.shabd.Setting;
+package com.dsciitp.shabd.quickActions;
 
 
 import android.os.Bundle;
@@ -10,15 +10,13 @@ import android.widget.TextView;
 
 import com.dsciitp.shabd.R;
 
-import java.util.zip.Inflater;
-
 /**
  * A simple {@link Fragment} subclass.
  */
-public class SettingFragment extends Fragment {
+public class QuickActionFragment extends Fragment {
 
 
-    public SettingFragment() {
+    public QuickActionFragment() {
         // Required empty public constructor
     }
 
@@ -26,9 +24,9 @@ public class SettingFragment extends Fragment {
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.fragment_setting, container, false);
-
-        return view;
+        TextView textView = new TextView(getActivity());
+        textView.setText(R.string.hello_blank_fragment);
+        return textView;
     }
 
 }
