+++ app/src/main/java/com/dsciitp/shabd/setting/SettingFragment.java
@@ -1,4 +1,4 @@
-package com.dsciitp.shabd.QuickActions;
+package com.dsciitp.shabd.setting;
 
 
 import android.os.Bundle;
@@ -6,17 +6,16 @@ import android.support.v4.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
-import android.widget.TextView;
 
 import com.dsciitp.shabd.R;
 
 /**
  * A simple {@link Fragment} subclass.
  */
-public class QuickActionFragment extends Fragment {
+public class SettingFragment extends Fragment {
 
 
-    public QuickActionFragment() {
+    public SettingFragment() {
         // Required empty public constructor
     }
 
@@ -24,9 +23,9 @@ public class QuickActionFragment extends Fragment {
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container,
                              Bundle savedInstanceState) {
-        TextView textView = new TextView(getActivity());
-        textView.setText(R.string.hello_blank_fragment);
-        return textView;
+        View view = inflater.inflate(R.layout.fragment_setting, container, false);
+
+        return view;
     }
 
 }
