+++ app/src/main/res/layout/activity_dictionary.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Dictionary.DictionaryActivity">
+    tools:context=".dictionary.DictionaryActivity">
     <RelativeLayout
         android:layout_width="fill_parent"
         android:layout_height="wrap_content"
