+++ app/src/main/res/layout/activity_learn.xml
@@ -4,6 +4,6 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".Learn.LearnActivity">
+    tools:context=".learn.LearnActivity">
 
 </android.support.constraint.ConstraintLayout>
