+++ app/src/main/res/layout/fragment_basic.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".BasicTopic.BasicFragment">
+    tools:context=".basicTopic.BasicFragment">
 
 
     <android.support.v7.widget.RecyclerView
